[react] Use global `React` for now
- Something is screwing up with browserify or react
- Put PostList in a separate component

- Need to figure this rubbish out!
