+++ client/js/app.js
@@ -1,7 +1,8 @@
 require('normalize-css');
-var React = require('react');
+React = require('react');
 
 var Post = require('./Post');
+var PostList = require('./PostList');
 var PostBox = require('./PostBox');
 
 React.render(<PostBox/>, document.querySelector('main'));
